fix: changed on behalf of item position
