+++ src/translations/screens/subscreens/PurchaseOverview.ts
@@ -210,7 +210,7 @@ const PurchaseOverviewTexts = {
     ),
     button: {
       payment: _('Til betaling', 'To payment', 'Til betaling'),
-      sendToOthers: _('Gå videre', 'Continue', 'Gå vidare'),
+      sendToOthers: _('Select recipient', 'Select recipient', 'Vel mottakar'),
     },
     ordinaryPriceA11yLabel: (priceString: string) =>
       _(
@@ -247,16 +247,16 @@ const PurchaseOverviewTexts = {
     ),
   },
   onBehalfOf: {
-    sectionTitle: _('Kjøp til andre', 'Buy for others', 'Kjøp til andre'),
     sectionSubText: _(
       'Den du kjøper billett til, må være innlogget i AtB-appen for å få billetten.',
-      'The person you buy a ticket for, must be logged in to the AtB app to get the ticket.',
+      'The person you buy a ticket for, must be logged in to the AtB to recieve the ticket.',
       'Den du kjøper billett til, må vere logga inn i AtB-appen for å få billetten.',
     ),
+
     sendToOthersText: _(
-      'Sendes til noen andre',
-      'Sending to someone else',
-      'Sendast til nokon andre',
+      'Send billetten til noen andre',
+      'Send the ticket to someone else',
+      'Send biletten til nokon andre',
     ),
   },
   ticketInformation: {
